requirements.txt
# Minimal requirements for the calibrated grid system
# Install with: pip install -r requirements-calibrated.txt

# Core image processing
Pillow>=11.0.0

# Computer vision for ArUco marker detection
opencv-python>=4.8.0

# NumPy is required by OpenCV
numpy>=1.24.0
